Rename auth files to user files
- Renamed authController.js to UserController.js
- Renamed authRoutes.js to UserRoutes.js
- Updated app.js to reflect these changes
